Added minix file transfers for the PT68K-2/4
I have added the ability to use this program with minix 1.4b that runs on the PT68K-2/4. This required changes to account for the differences in the file stat structure between UniFLEX and minix. A new check box has been added to inform the application that the user is connected to a minix box instead of a UniFLEX box. I also cleaned up some of the text box populations when browsing the local file system and the remote file system.