fixed the file stat handling
fixed the file stat handling for difference between UniFLEX and Minix stat structure. It now uses a common structure on the PC end that can accomodate both structures instead of having a different structure for each. The differences are now handled in the function that fills the structure from the raw bytes read.