Added my current SD Card image for booting minix on PT68K-2/4
I have added the Win 32 Disk Imager generated copy of the SD Card that I use in my PT68K-2/4 machines to the project. The SD Card I use is only 256MB since that is plenty of room to do development work and it is quick to backup and restore if something goes wrong between the keyboard and the chair (PEBKAC). Just put it in the SD Card adapter attached to the PT-XT IDE interface and type either WD or MI at the HUMBUG prompt. Once the boot finishes, type root for the user at the login prompt. The password is merlin.