Added internal editor
Added the same internal editor that is in FloppyMaintenance. Removed the FLEX specific code and made it more generic. It still has all of the same features like Search (Find), Line Numbers (on/off), goto line save and save as. It does not yet implement auto saving back to the remote. That will come later and will require some VERY extensive testing. The search function will center the line it finds the text on if it is not already visible on the page. It it is already visible on the page, it will just position the cursor to it. The status line will reflect the position of the cursor.